Check sum

+++ cz4013_server/src/com/company/cz4013/util/AdlerCheckSum.java
@@ -6,10 +6,10 @@ import java.util.zip.Checksum;
 public class AdlerCheckSum {
 
 
-    public static int checkSum(byte[] input){
+    public static long checkSum(byte[] input){
         Checksum checksumEngine = new Adler32();
         checksumEngine.update(input, 0, input.length);
-        int checksum = (int)checksumEngine.getValue();
+        long checksum = checksumEngine.getValue();
         System.out.println(checksum);
         return checksum;
     }
